makes IPGeo service configurable
